requirements.txt
streamlit
gspread
google-auth
pandas
python-dotenv
openai
json5
